Refactor content components
