+++ src/components/content-views/VideoListRenderItem.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import { View, Image, TouchableOpacity, StyleSheet, Text } from 'react-native';
+import { Image, TouchableOpacity, Text } from 'react-native';
 import { faPlay } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
+import { styles } from './index.styles';
 
 const VideoListRenderItem = ({ item, navigation }) => {
   return (
@@ -19,29 +20,5 @@ const VideoListRenderItem = ({ item, navigation }) => {
     </TouchableOpacity>
   );
 };
-
-const styles = StyleSheet.create({
-  itemContainer: {
-    margin: 20,
-    alignItems: 'center',
-    borderBottomWidth: 1,
-    paddingBottom: 20,
-    borderColor: 'grey',
-  },
-  itemIcon: {
-    position: 'absolute',
-    top: '45%',
-    left: '45%',
-  },
-  itemImage: {
-    height: 200,
-    width: '100%',
-    marginBottom: 20,
-  },
-  title: {
-    fontSize: 16,
-    fontWeight: '500',
-  },
-});
 
 export default VideoListRenderItem;
