+++ src/components/content-views/index.js
@@ -1,0 +1,5 @@
+import TodayReadRenderItem from './TodayReadRenderItem';
+import VideoListRenderItem from './VideoListRenderItem';
+import WatchListRenderItem from './WatchListRenderItem';
+
+export { TodayReadRenderItem, VideoListRenderItem, WatchListRenderItem };
