+++ src/components/content/ArticleCard.styles.js
@@ -1,37 +1,8 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-import {
-  View,
-  StyleSheet,
-  Text,
-  Image,
-  Dimensions,
-  TouchableOpacity,
-} from 'react-native';
+import { StyleSheet, Dimensions } from 'react-native';
 
 const { width, height } = Dimensions.get('window');
 
-const ArticleCard = ({ item, navigation }) => {
-  return (
-    <View style={styles.cardView}>
-      <TouchableOpacity
-        onPress={() => navigation.navigate('Article Page', { id: item._id })}>
-        <Text style={styles.title}> {item.title}</Text>
-        <Text style={styles.author}>{item.author} </Text>
-        <Text style={styles.violence_type}>Tags: {item.violence_type}</Text>
-        <Image
-          style={styles.image}
-          source={item.url_to_image ? { uri: item.url_to_image } : null}
-        />
-        <Text style={styles.text}>{item.text.substring(0, 150)} ...</Text>
-        <Text style={styles.created_at}>Date: {item.created_at}</Text>
-      </TouchableOpacity>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
+export const styles = StyleSheet.create({
   cardView: {
     backgroundColor: 'white',
     margin: width * 0.03,
@@ -88,9 +59,3 @@ const styles = StyleSheet.create({
     marginTop: -8,
   },
 });
-
-export default ArticleCard;
-
-ArticleCard.propTypes = {
-  item: PropTypes.object.isRequired,
-};
